fix: workaround because custom useQuery doesn't immediately update the data
This is a workaround for a weird bug: When changing to amino acids, the plot crashes.
On the first update run, the old data for nucleotides would still be returned by useQuery.
Only on the second run, it returns the new data.

Also make the unit test independent of local time.